fixed comments, format & removed unused functions
